# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the Green Ecolution Management System.",
        "title": "Green Space Management API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Green Ecolution",
            "url": "https://green-ecolution.de",
            "email": "info@green-ecolution.de"
        },
        "license": {
            "name": "AGPL",
            "url": "https://raw.githubusercontent.com/green-ecolution/green-ecolution/develop/LICENSE"
        },
        "version": "develop"
    },
    "paths": {
...
        "/v1/tree": {
            "get": {
                "security": [
                    {
                        "Keycloak": []
                    }
                ],
                "description": "Get all trees",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tree"
                ],
                "summary": "Get all trees",
                "operationId": "get-all-trees",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Provider",
                        "name": "provider",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "watering status (good, moderate, bad)",
                        "name": "watering_statuses",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "planting_years",
                        "name": "planting_years",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "has cluster",
                        "name": "has_cluster",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/TreeList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Keycloak": []
                    }
                ],
                "description": "Create tree",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tree"
                ],
                "summary": "Create tree",
                "operationId": "create-tree",
                "parameters": [
                    {
                        "description": "Tree to create",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/TreeCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Tree"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            }
        },
        "/v1/tree/sensor/{sensor_id}": {
            "get": {
                "security": [
                    {
                        "Keycloak": []
                    }
                ],
                "description": "Get tree by sensor ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tree Sensor"
                ],
                "summary": "Get tree by sensor ID",
                "operationId": "get-tree-by-sensor-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sensor ID",
                        "name": "sensor_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Tree"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            }
        },
        "/v1/tree/{tree_id}": {
            "get": {
                "security": [
                    {
                        "Keycloak": []
                    }
                ],
                "description": "Get tree by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tree"
                ],
                "summary": "Get tree by ID",
                "operationId": "get-trees",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tree ID",
                        "name": "tree_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Tree"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Keycloak": []
                    }
                ],
                "description": "Update tree",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tree"
                ],
                "summary": "Update tree",
                "operationId": "update-tree",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tree ID",
                        "name": "tree_id",
                        "in": "path"
                    },
                    {
                        "description": "Tree to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/TreeUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Tree"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Keycloak": []
                    }
                ],
                "description": "Delete tree",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tree"
                ],
                "summary": "Delete tree",
                "operationId": "delete-tree",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tree ID",
                        "name": "tree_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "HTTPError": {
            "type": "object",
            "required": [
                "code",
                "error",
                "method",
                "path"
            ],
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "method": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
...
        "Pagination": {
            "type": "object",
            "required": [
                "current_page",
                "next_page",
                "prev_page",
                "total_pages",
                "total_records"
            ],
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "next_page": {
                    "type": "integer"
                },
                "prev_page": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                },
                "total_records": {
                    "type": "integer"
                }
            }
        },
...
        "Sensor": {
            "type": "object",
            "required": [
                "created_at",
                "id",
                "latest_data",
                "latitude",
                "longitude",
                "status",
                "updated_at"
            ],
            "properties": {
                "additional_information": {
                    "type": "object",
                    "additionalProperties": true
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "latest_data": {
                    "$ref": "#/definitions/SensorData"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "provider": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/SensorStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "SensorData": {
            "type": "object",
            "required": [
                "battery",
                "created_at",
                "humidity",
                "temperature",
                "updated_at",
                "watermarks"
            ],
            "properties": {
                "battery": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "humidity": {
                    "type": "number"
                },
                "temperature": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "watermarks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/WatermarkResponse"
                    }
                }
            }
        },
...
        "SensorStatus": {
            "type": "string",
            "enum": [
                "online",
                "offline",
                "unknown"
            ],
            "x-enum-varnames": [
                "SensorStatusOnline",
                "SensorStatusOffline",
                "SensorStatusUnknown"
            ]
        },
...
        "Tree": {
            "type": "object",
            "required": [
                "created_at",
                "description",
                "id",
                "latitude",
                "longitude",
                "number",
                "planting_year",
                "species",
                "updated_at",
                "watering_status"
            ],
            "properties": {
                "additional_information": {
                    "type": "object",
                    "additionalProperties": true
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_watered": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "number": {
                    "type": "string"
                },
                "planting_year": {
                    "type": "integer"
                },
                "provider": {
                    "type": "string"
                },
                "sensor": {
                    "$ref": "#/definitions/Sensor"
                },
                "species": {
                    "type": "string"
                },
                "tree_cluster_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "watering_status": {
                    "$ref": "#/definitions/WateringStatus"
                }
            }
        },
...
        "TreeCreate": {
            "type": "object",
            "required": [
                "description",
                "latitude",
                "longitude",
                "number",
                "planting_year",
                "species"
            ],
            "properties": {
                "additional_information": {
                    "type": "object",
                    "additionalProperties": true
                },
                "description": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "number": {
                    "type": "string"
                },
                "planting_year": {
                    "type": "integer"
                },
                "provider": {
                    "type": "string"
                },
                "sensor_id": {
                    "type": "string"
                },
                "species": {
                    "type": "string"
                },
                "tree_cluster_id": {
                    "type": "integer"
                }
            }
        },
        "TreeList": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Tree"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/Pagination"
                }
            }
        },
        "TreeUpdate": {
            "type": "object",
            "required": [
                "description",
                "latitude",
                "longitude",
                "number",
                "planting_year",
                "species"
            ],
            "properties": {
                "additional_information": {
                    "type": "object",
                    "additionalProperties": true
                },
                "description": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "number": {
                    "type": "string"
                },
                "planting_year": {
                    "type": "integer"
                },
                "provider": {
                    "type": "string"
                },
                "sensor_id": {
                    "type": "string"
                },
                "species": {
                    "type": "string"
                },
                "tree_cluster_id": {
                    "type": "integer"
                }
            }
        },
...
        "WateringStatus": {
            "type": "string",
            "enum": [
                "good",
                "moderate",
                "bad",
                "just watered",
                "unknown"
            ],
            "x-enum-varnames": [
                "WateringStatusGood",
                "WateringStatusModerate",
                "WateringStatusBad",
                "WateringStatusJustWatered",
                "WateringStatusUnknown"
            ]
        },
        "WatermarkResponse": {
            "type": "object",
            "required": [
                "centibar",
                "depth",
                "resistance"
            ],
            "properties": {
                "centibar": {
                    "type": "integer"
                },
                "depth": {
                    "type": "integer"
                },
                "resistance": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "Keycloak": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://auth.green-ecolution.de/realms/green-ecolution-dev/protocol/openid-connect/auth",
            "tokenUrl": "https://auth.green-ecolution.de/realms/green-ecolution-dev/protocol/openid-connect/token"
        }
    }
}